完成user数据表“utype”字段添加

+++ src/main/java/com/grad/dao/UserMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Mapper;
 public interface UserMapper {
 
     public void addUser(User user);
-    public User selectUserById(int uid);
-    public int selectMaxUid();
+    public User selectUserById(long uid);
+    public long selectMaxUid();
 
 }
